refactor: consolidate config, deduplicate code, and split monolithic modules
