+++ app/model_analysis/__init__.py
@@ -1,11 +1,11 @@
-from model_analysis.charts import (
+from app.model_analysis.charts import (
     build_loss_chart,
     build_precision_recall_chart,
     build_val_map50_chart,
     build_val_map5095_chart,
 )
-from model_analysis.glossary import METRIC_DEFINITIONS, glossary_rows
-from model_analysis.presenters import build_kpi_items, format_run_context
+from app.model_analysis.glossary import METRIC_DEFINITIONS, glossary_rows
+from app.model_analysis.presenters import build_kpi_items, format_run_context
 
 __all__ = [
     "METRIC_DEFINITIONS",
