+++ pyproject.toml
@@ -19,8 +19,8 @@ dev = [
 ]
 
 [tool.setuptools.packages.find]
-where = ["src"]
-include = ["murawa*"]
+where = ["src", "."]
+include = ["murawa*", "app*"]
 
 [tool.pytest.ini_options]
 minversion = "8.0"
