+++ scripts/train.py
@@ -5,14 +5,12 @@ import logging
 from datetime import datetime, timezone
 from pathlib import Path
 
-from murawa.settings import PROJECT_ROOT
+from murawa.settings import MODELS_CHECKPOINTS, MODELS_METADATA, PROJECT_ROOT
 from murawa.data import DataLoaderError, load_training_split, summarize_variant
 from murawa.models import build_training_adapter, normalize_model_name
 from murawa.services.runtime.artifacts import (
     ArtifactManifest,
     ArtifactWriteContext,
-    CKPT_DIR,
-    META_DIR,
     StandardizedArtifactCallback,
     make_run_name,
     save_config,
@@ -80,8 +78,8 @@ def main() -> int:
     created_at = datetime.now(timezone.utc)
     run_name = make_run_name(model_name, args.dataset_variant, created_at, tag=run_tag)
 
-    ckpt_dir = ROOT / CKPT_DIR / run_name
-    meta_dir = ROOT / META_DIR / run_name
+    ckpt_dir = ROOT / MODELS_CHECKPOINTS / run_name
+    meta_dir = ROOT / MODELS_METADATA / run_name
     if _non_empty_dir(ckpt_dir) or _non_empty_dir(meta_dir):
         logger.error(
             "Run directory already exists and is not empty: %s or %s. "
