+++ src/murawa/settings.py
@@ -40,12 +40,6 @@ MODELS_METADATA = _path("models_metadata")
 OUTPUTS_PREDICTIONS = _path("outputs_predictions")
 OUTPUTS_VIDEOS = _path("outputs_videos")
 
-# Backward-compatible aliases used across the codebase during migration.
-READY_ROOT = DATA_READY
-PREDICTIONS_ROOT = OUTPUTS_PREDICTIONS
-CKPT_DIR = MODELS_CHECKPOINTS
-META_DIR = MODELS_METADATA
-
 _dataset = _load_project_config().get("dataset", {})
 if isinstance(_dataset, dict):
     _default_variant = _dataset.get("default_variant")
@@ -72,10 +66,41 @@ if isinstance(_dataset, dict):
         if isinstance(_video_suffixes, list) and _video_suffixes
         else {".mp4", ".avi", ".mov", ".mkv", ".webm"}
     )
+    _bootstrap = _dataset.get("bootstrap")
 else:
     SPLITS = ("train", "valid", "test")
     IMAGE_SUFFIXES = {".jpg", ".jpeg", ".png", ".bmp"}
     VIDEO_SUFFIXES = {".mp4", ".avi", ".mov", ".mkv", ".webm"}
+    _bootstrap = None
+
+if isinstance(_bootstrap, dict):
+    DEFAULT_FRAME_STEP = int(_bootstrap.get("frame_step", 30))
+    _split_ratios = _bootstrap.get("split_ratios", {})
+    BOOTSTRAP_SPLIT_RATIOS: dict[str, float] = (
+        {
+            "train": float(_split_ratios.get("train", 0.8)),
+            "valid": float(_split_ratios.get("valid", 0.1)),
+            "test": float(_split_ratios.get("test", 0.1)),
+        }
+        if isinstance(_split_ratios, dict)
+        else {"train": 0.8, "valid": 0.1, "test": 0.1}
+    )
+    BOOTSTRAP_VALID_FROM_TRAIN_RATIO = float(_bootstrap.get("valid_from_train_ratio", 0.1))
+    _default_resolution = _bootstrap.get("default_resolution", {})
+    if isinstance(_default_resolution, dict):
+        DEFAULT_FRAME_WIDTH = int(_default_resolution.get("width", 1920))
+        DEFAULT_FRAME_HEIGHT = int(_default_resolution.get("height", 1080))
+    else:
+        DEFAULT_FRAME_WIDTH = 1920
+        DEFAULT_FRAME_HEIGHT = 1080
+    COLOR_CLUSTER_ITERATIONS = int(_bootstrap.get("color_cluster_iterations", 12))
+else:
+    DEFAULT_FRAME_STEP = 30
+    BOOTSTRAP_SPLIT_RATIOS = {"train": 0.8, "valid": 0.1, "test": 0.1}
+    BOOTSTRAP_VALID_FROM_TRAIN_RATIO = 0.1
+    DEFAULT_FRAME_WIDTH = 1920
+    DEFAULT_FRAME_HEIGHT = 1080
+    COLOR_CLUSTER_ITERATIONS = 12
 
 SUPPORTED_IMAGE_EXTENSIONS = IMAGE_SUFFIXES
 
@@ -103,6 +128,20 @@ def resolve_project_root(project_root: Path | None = None) -> Path:
     if project_root is not None:
         return project_root.resolve()
     return PROJECT_ROOT.resolve()
+
+
+def infer_project_root_from_output_dir(output_dir: Path) -> Path:
+    """Infer project root from a checkpoint directory (<root>/models/checkpoints/<run>)."""
+    output_dir = output_dir.resolve()
+    if len(output_dir.parents) >= 3:
+        return output_dir.parents[2]
+    return PACKAGE_ROOT.parents[1]
+
+
+def infer_project_root_from_ready_root(ready_root: Path) -> Path:
+    """Infer project root from data/ready path (<root>/data/ready)."""
+    ready_root = Path(ready_root).resolve()
+    return ready_root.parents[1]
 
 
 def resolve_variant_dir(
